% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_type_annotation.R
\name{cell_type_annotation}
\alias{cell_type_annotation}
\title{Run mixtures through the NMF model to get the cell type composition.}
\usage{
cell_type_annotation(
  nmf_mod,
  mixture_transcriptome,
  transf,
  reference_profiles,
  st_position = NULL,
  min_cont = 0.01,
  regr = "lar",
  regr_type = "nnls"
)
}
\arguments{
\item{nmf_mod}{Object of class NMF containing the trained nNMF model.}

\item{mixture_transcriptome}{Matrix of dimensions GENES x SPOTS containing the spatial transcriptomics data.}

\item{transf}{Character string specifying the transformation to normalize the count matrix: "uv" (unit variance) or "raw" (no transformation).}

\item{reference_profiles}{Matrix of dimensions TOPICS x CELLS containing the cell type profiles.}

\item{st_position}{Optional. Matrix containing the spatial coordinates of the spots. If provided, it should have two columns: "x" and "y".}

\item{min_cont}{Numeric value indicating the minimum contribution expected from a cell in that spot. Default is 0.01, meaning that cell types with a weight coefficient of at least 1\% of the total will be accepted.}

\item{regr}{Character string specifying the regression method for deconvolution: "lar" (Least Angle Regression) or "nnls" (Non-negative Least Squares). Default is "lar".}

\item{regr_type}{Character string specifying the type of regression: "nnls", "lar", or "none". Default is "nnls".}
}
\value{
A matrix with the coefficients of the spatial mixtures, representing the inferred cell type proportions for each spot.
}
\description{
Run mixtures through the NMF model to get the cell type composition.
}
\examples{
# Example usage:
# nmf_mod, mixture_transcriptome, reference_profiles, and st_position are example datasets
result <- cell_type_annotation(nmf_mod = nmf_mod,
                              mixture_transcriptome = mixture_transcriptome,
                              transf = "uv",
                              reference_profiles = reference_profiles,
                              st_position = st_position,
                              min_cont = 0.01,
                              regr = "lar",
                              regr_type = "nnls")

}
