% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_label.R
\name{dist_calc}
\alias{dist_calc}
\title{Calculate Distances Between Matrices}
\usage{
dist_calc(x, y)
}
\arguments{
\item{x}{First matrix.}

\item{y}{Second matrix.}
}
\value{
A vector of distances.
}
\description{
This function calculates the distances between corresponding rows of two matrices.
}
\examples{
# Example usage
x <- matrix(1:6, ncol = 2)
y <- matrix(7:12, ncol = 2)
dists <- dist_calc(x, y)
print(dists)
}
