% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_method.R
\name{.getTensor}
\alias{.getTensor}
\title{Convert a matrix to a PyTorch tensor}
\usage{
.getTensor(m, device)
}
\arguments{
\item{m}{Matrix to convert}

\item{device}{Device to place the tensor on (e.g., "cpu" or "cuda")}
}
\value{
PyTorch tensor
}
\description{
Convert a matrix to a PyTorch tensor
}
\examples{
Xt <- .getTensor(X, device)
}
