% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_performances.R
\name{each_class_performances}
\alias{each_class_performances}
\title{Test the performance of the model for each class on synthetic generated test spots.}
\usage{
each_class_performances(test_spots_metadata_mtrx, spot_composition_mtrx)
}
\arguments{
\item{test_spots_metadata_mtrx}{Object of class matrix containing the ground truth composition of each spot as obtained from the function syn_spot_comb_topic_fun.R, 2nd element.}

\item{spot_composition_mtrx}{Object of class matrix with the predicted topic probability distributions for each spot.}
}
\value{
This function returns a list with performance metrics for each class.
}
\description{
Test the performance of the model for each class on synthetic generated test spots.
}
\examples{
class_performance_metrics <- each_class_performances(test_spots_metadata_mtrx, spot_composition_mtrx)
print(class_performance_metrics)
}
