% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_label.R
\name{get_top_k_intersection}
\alias{get_top_k_intersection}
\title{Get Top k Intersections}
\usage{
get_top_k_intersection(knn.indexes, knn.diff, k)
}
\arguments{
\item{knn.indexes}{Matrix of k-nearest neighbor indices.}

\item{knn.diff}{Matrix of k-nearest neighbor indices based on differences.}

\item{k}{Number of nearest neighbors to consider.}
}
\value{
A list of intersections for each row.
}
\description{
This function gets the intersection of the top k indices from two sets of k-nearest neighbors.
}
\examples{
# Example usage
knn.indexes <- matrix(1:12, ncol = 4)
knn.diff <- matrix(13:24, ncol = 4)
intersections <- get_top_k_intersection(knn.indexes, knn.diff, k = 3)
print(intersections)
}
