% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installPython.R
\name{installPython}
\alias{installPython}
\title{Install Python and necessary packages using Miniconda}
\usage{
installPython(
  conda = "auto",
  python.version = "3.8",
  install.conda = FALSE,
  miniconda.path = NULL
)
}
\arguments{
\item{conda}{Path to the Conda executable or "auto" to automatically detect it.}

\item{python.version}{Version of Python to install (default is "3.8").}

\item{install.conda}{Logical indicating whether to install Miniconda if not detected (default is FALSE).}

\item{miniconda.path}{Path to the Miniconda installation directory (default is NULL, which uses the default path).}
}
\value{
None
}
\description{
This function ensures that the required Python environment and packages are installed.
}
\examples{
installPython()
}
