% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_label.R
\name{rescale_function}
\alias{rescale_function}
\title{Rescale Values}
\usage{
rescale_function(x, to = c(0, 1), from = range(x, na.rm = TRUE, finite = TRUE))
}
\arguments{
\item{x}{Vector of values to rescale.}

\item{to}{Range to rescale to (default is c(0, 1)).}

\item{from}{Range to rescale from (default is the range of x).}
}
\value{
Rescaled values.
}
\description{
This function rescales values to a specified range.
}
\examples{
# Example usage
x <- c(1, 2, 3, 4, 5)
rescaled_x <- rescale_function(x, to = c(0, 10))
print(rescaled_x)
}
