% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_performances.R
\name{test_performances}
\alias{test_performances}
\title{Test the performance of the model on synthetic generated test spots.}
\usage{
test_performances(test_spots_metadata_mtrx, spot_composition_mtrx)
}
\arguments{
\item{test_spots_metadata_mtrx}{Object of class matrix containing the ground truth composition of each spot as obtained from the function syn_spot_comb_topic_fun.R, 2nd element.}

\item{spot_composition_mtrx}{Object of class matrix with the predicted topic probability distributions for each spot.}
}
\value{
This function returns a list with TP, TN, FP, FN, and the Jensen-Shannon Divergence index.
}
\description{
Test the performance of the model on synthetic generated test spots.
}
\examples{
# Example usage
# Assuming test_spots_metadata_mtrx and spot_composition_mtrx are already defined
performance_metrics <- test_performances(test_spots_metadata_mtrx, spot_composition_mtrx)
print(performance_metrics)
}
