% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_distribution_analysis.R
\name{topic_distribution_analysis}
\alias{topic_distribution_analysis}
\title{Run test spots through the basis to get the pertinent coefficients. To do this for every spot we are going to set up a system of linear equations where we need to find the coefficient, we will use non-negative least squares to determine the best coefficient fit. Returns a matrix of KxnSPOTS dimensions.}
\usage{
topic_distribution_analysis(
  nmf_mod,
  mixture_transcriptome,
  transf = "uv",
  regr = "lar"
)
}
\arguments{
\item{nmf_mod}{Object of class NMFModel obtained from the NMF package.}

\item{mixture_transcriptome}{Matrix of dimensions GENESxSPOTS.}

\item{transf}{Transformation to normalize the count matrix: "cpm" (Counts per million), "uv" (unit variance), "raw" (no transformation applied).}

\item{regr}{Regression method to use: "nnls" (non-negative least squares), "lar" (least angle regression).}
}
\value{
This function returns a matrix with the coefficients of the spatial mixtures.
}
\description{
Run test spots through the basis to get the pertinent coefficients. To do this for every spot we are going to set up a system of linear equations where we need to find the coefficient, we will use non-negative least squares to determine the best coefficient fit. Returns a matrix of KxnSPOTS dimensions.
}
\examples{
# Example usage
# Assuming nmf_mod and mixture_transcriptome are already defined
coef_matrix <- topic_distribution_analysis(nmf_mod, mixture_transcriptome, transf = "uv", regr = "nnls")
print(coef_matrix)
}
