% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_identification.R
\name{topic_identification}
\alias{topic_identification}
\title{This function carries out the ssNMF and returns an NMF object.}
\usage{
topic_identification(
  se_sc,
  mtrx_spatial,
  cluster_markers,
  clust_vr,
  ntop = NULL,
  transf = "uv",
  nmf = "ssNMF",
  numiters = 100,
  lam = 1,
  assay = "RNA",
  slot = "counts"
)
}
\arguments{
\item{se_sc}{Object of class Seurat with the scRNAseq data.}

\item{mtrx_spatial}{Object of class matrix of shape GENESxSPOT.}

\item{cluster_markers}{Object of class dataframe obtained from the function Seurat::FindAllMarkers().}

\item{clust_vr}{Object of class character; Name of the variable containing the cell clustering.}

\item{ntop}{Object of class "numeric" or NULL; number of unique markers per cluster used to seed the model, by default NULL. If NULL it uses all of them.}

\item{transf}{Transformation to normalize the count matrix: "uv" (unit variance), "raw" (no transformation applied). By default "uv".}

\item{nmf}{Object of class character; Type of method to use to find W and H. Look at NMF package for the options and specifications, by default "ssNMF".}

\item{numiters}{Number of iterations for the NMF algorithm, by default 100.}

\item{lam}{Regularization parameter for the NMF algorithm, by default 1.}

\item{assay}{Object of class character; From which assay to grab the expression data to train the model, by default "RNA".}

\item{slot}{Object of class character; From which slot to grab the expression data to train the model, by default "counts".}
}
\value{
This function returns a list with the initialized matrices H and W.
}
\description{
This function carries out the ssNMF and returns an NMF object.
}
\examples{
# Example usage
# Assuming se_sc, mtrx_spatial, and cluster_markers are already defined
nmf_result <- topic_identification(se_sc, mtrx_spatial, cluster_markers, clust_vr = "seurat_clusters")
print(nmf_result)
}
